+++ fast-toolkit/src/main/java/com/thankjava/toolkit/bean/encrypit/RSASignAlgorithm.java
@@ -1,0 +1,7 @@
+package com.thankjava.toolkit.bean.encrypit;
+
+public enum RSASignAlgorithm {
+
+    SHA1withRSA,
+    SHA256withRSA
+}